Separated Extensions.cs methods into their own files
