+++ src/Bot/Extensions/EnumerableExtension.cs
@@ -1,0 +1,7 @@
+﻿namespace OPZBot.Extensions;
+
+public static class EnumerableExtension
+{
+    public static IEnumerable<TSource> ExcludeFirst<TSource>(this IEnumerable<TSource> source) where TSource : class
+        => source.Where(x => x != source.First());
+}
